requirements.txt
requests
pandas
beautifulsoup4
python-dotenv
python-telegram-bot
pytz
nest-asyncio
